auth backend and frontend (80%)

+++ app/views/post/index.php
@@ -1,6 +1,6 @@
 <div class="shadow-lg px-4 py-3 rounded">
     <div class="d-flex justify-content-between mb-3">
-        <h4>Table Posts</h4>
+        <h4>Table Post</h4>
         <a href="<?= BASE_URL ?>/post/create" class="btn btn-primary">Add Post</a>
     </div>
     <table class="table">
@@ -16,20 +16,26 @@
         </tr>
     </thead>
     <tbody>
-        <?php foreach($data['posts'] as $index => $post) : ?>
+        <?php if(count($data['posts']) > 0) : ?>
+            <?php foreach($data['posts'] as $index => $post) : ?>
+                <tr>
+                    <td><?= ++$index ?></td>
+                    <td><?= $post['title'] ?></td>
+                    <td><?= $post['category_name'] ?></td>
+                    <td><?= $post['author'] ?></td>
+                    <td><?= date_format(date_create($post['created_at']), "d M Y H:i:s"); ?></td>
+                    <td><?= date_format(date_create($post['updated_at']), "d M Y H:i:s"); ?></td>
+                    <td>
+                        <a href="<?= BASE_URL ?>/post/edit/<?= $post['id'] ?>" class="btn btn-warning btn-sm me-2">Edit</a>
+                        <a href="<?= BASE_URL ?>/post/destroy/<?= $post['id'] ?>" onclick="return confirm('Are you sure you want to delete this post?')" class="btn btn-danger btn-sm">Delete</a>
+                    </td>
+                </tr>
+            <?php endforeach; ?>
+        <?php else : ?>
             <tr>
-                <td><?= ++$index ?></td>
-                <td><?= $post['title'] ?></td>
-                <td><?= $post['category_name'] ?></td>
-                <td><?= $post['author'] ?></td>
-                <td><?= date_format(date_create($post['created_at']), "d M Y H:i:s"); ?></td>
-                <td><?= date_format(date_create($post['updated_at']), "d M Y H:i:s"); ?></td>
-                <td>
-                    <a href="<?= BASE_URL ?>/post/edit/<?= $post['id'] ?>" class="btn btn-warning btn-sm me-2">Edit</a>
-                    <a href="<?= BASE_URL ?>/post/destroy/<?= $post['id'] ?>" onclick="return confirm('Are you sure you want to delete this post?')" class="btn btn-danger btn-sm">Delete</a>
-                </td>
+                <td colspan="7" class="text-center">Data post not found</td>
             </tr>
-        <?php endforeach; ?>
+        <?php endif; ?>
     </tbody>
     </table>
 </div>
